TODO: Parser

+++ includes/graphics.h
@@ -29,7 +29,7 @@ typedef struct		s_bitmap
 }					t_bitmap;
 
 void			draw_screen(t_data *data);
-void			draw_pix(t_data *data, double x, double y, int color);
+void			draw_pix(t_data *data, float x, float y, int color);
 void			draw_line(t_data *data, t_vec2 a, t_vec2 b, t_vec2 colors);
 void			draw_line_x(t_data *data, t_vec2 a, t_vec2 b, t_vec2 colors);
 void			draw_line_y(t_data *data, t_vec2 a, t_vec2 b, t_vec2 colors);
@@ -37,7 +37,7 @@ t_bitmap		*new_bitmap(t_data *data, int width, int height);
 void			bitmap_draw_pix(t_bitmap *b, int x, int y, int color);
 void			bitmap_update(t_bitmap *b);
 void			clear_bitmap(t_bitmap *b);
-int				col_lerp(int a, int b, double n);
+int				col_lerp(int a, int b, float n);
 int				color(int r, int g, int b);
 int				color_angle(int angle, int sat);
 int				to_color(t_vec3 v);
